+++ backend/models/user_role.py
@@ -1,5 +1,5 @@
 # --- third party imports ---
-from sqlalchemy import Integer, ForeignKey
+from sqlalchemy import BigInteger, ForeignKey
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 # --- local imports ---
@@ -13,14 +13,14 @@ class UserRole(Base):
     __tablename__ = "user_role"
 
     user_id: Mapped[int] = mapped_column(
-        Integer,
+        BigInteger,
         ForeignKey(User.id, ondelete="CASCADE"),
         nullable=False,
         index=True,
         primary_key=True,
     )
     role_id: Mapped[int] = mapped_column(
-        Integer,
+        BigInteger,
         ForeignKey(Role.id, ondelete="CASCADE"),
         nullable=False,
         index=True,
