+++ backend/schemas/appointments.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 from pydantic import BaseModel, ConfigDict
-from sqlalchemy import Integer
+from sqlalchemy import BigInteger
 
 from models.appointments import AppointmentStatus, Appointments
 
@@ -8,15 +8,15 @@ from models.appointments import AppointmentStatus, Appointments
 class AppointmentsPublic(BaseModel):
     datetime_start: datetime
     datetime_end: datetime
-    room_id: Integer
+    room_id: BigInteger
 
 
 class AppointmenstSchema(AppointmentsPublic):
-    id: Integer
+    id: BigInteger
     created_at: datetime
     reason: str
     status: AppointmentStatus
-    applicant_id: Integer
-    approver_id: Integer
+    applicant_id: BigInteger
+    approver_id: BigInteger
 
     model_config = ConfigDict(from_attributes=True)
